+++ release_notes.md
@@ -4,3 +4,5 @@
 
 - Added the `EVENT_ENQUEUED` event when a task has been enqueued.
 - Added support for [huey-exporter](https://github.com/APGSGA/huey-exporter).
+- Rename `fire_enqueued_event` to `fire_enqueued_events`.
+- Moved `fire_enqueued_events` to huey config.